requirements.txt
psutil>=5.8.0
watchdog>=2.2.0
pandas>=1.3.0
matplotlib>=3.4.0
schedule>=1.1.8
pydantic>=1.8.0
PyYAML>=5.4.0
keyring>=23.0.0
python-dotenv>=0.19.0
